Fix various places where Compaq's cxx compiler is overly strict (it doesn't export some symbols from std:: into global namespace, even though they come from system headers.
git-svn-id: https://svn.dealii.org/trunk@4784 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/base/source/tensor.cc
@@ -42,8 +42,8 @@ template <int rank_, int dim>
 void
 Tensor<rank_, dim>::unroll( Vector<double>& result) const
 {
-  Assert(result.size()==pow(dim,rank_),
-	 ExcWrongVectorSize(static_cast<unsigned int>(pow(dim,rank_)),
+  Assert(result.size()==std::pow(static_cast<double>(dim),rank_),
+	 ExcWrongVectorSize(static_cast<unsigned int>(std::pow(static_cast<double>(dim),rank_)),
 			    result.size()));
 
   unsigned index = 0;
